Edit function createDecimal and function output
Conditionals for function output was edited to check for zeros.
Zeros need to be registered like any other number except when
the number is itself a zero. You do not want repeating zeros.

Conditionals for createDecimal was also edited to make sure that num1
and num2 can equal zero as long as the zero does not repeat.